requirements.txt
Django==2.2.10
django-crispy-forms==1.9.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2-binary==2.9.2
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
whitenoise==5.1.0
